Move FormulaAppCompatActivity into android package.
