+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Move `ActivityStoreContext` into `android` package.
 - Move `ActivityStore` into `android` package.
 - Move `FragmentId` into `android` package.
+- Move `FormulaAppCompatActivity` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
