+++ formula-android-tests/src/test/java/com/instacart/formula/ActivityLifecycleEventTest.kt
@@ -6,7 +6,7 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.instacart.formula.integration.FormulaAppCompatActivity
+import com.instacart.formula.android.FormulaAppCompatActivity
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
